fix(wizard): move tooltip
