+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/shared/AgentForm/EnvironmentUrlField.tsx
@@ -9,11 +9,12 @@ export function EnvironmentUrlField() {
   return (
     <FormControl
       label="Environment URL"
-      tooltip="URL or IP address of a Docker host. The Docker API must be exposed over a TCP port. Please refer to the Docker documentation to configure it."
       errors={meta.error}
       required
+      inputId="environment-url-field"
     >
       <Field
+        id="environment-url-field"
         name="environmentUrl"
         as={Input}
         placeholder="e.g. 10.0.0.10:9001 or tasks.portainer_agent:9001"
