+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/shared/AgentForm/index.tsx
@@ -1,0 +1,1 @@
+export { AgentForm } from './AgentForm';
